refactor: header 컴포넌트 디렉토리 위치 변경
FSD 구조에 맞게 ui로 위치 변경
